+++ CHANGELOG.md
@@ -1,6 +1,27 @@
 # Change Log
 All notable changes to this project will be documented in this file. 
 
+## 0.2.3 - 2016-10-01
+
+### Changed
+
+* Removed `:list {:indent-body n}`, made it `:indent`.  Made `:indent-arg`
+optional, uses `:indent` if it doesn't appear.
+* Added `:support` keyword (replaced undocumented `:explain-all`).
+* Added documentation for `:justify?` in readme.
+* Added `:key-ignore` and `:key-ignore-silent` for `:map` formatting.
+* Added `:to-string?` for `:record` formatting.
+
+### Removed
+
+* Removed (get-options) from zprint.core, took need for it out of readme.
+
+### Fixed
+
+* A problem where the last newline in a file was getting lost.
+* A problem where a comment inside of a fn starting in col 0 was getting
+an extra newline.
+
 ## 0.2.2 - 2016-09-24
 ### Changed